requirements.txt
flask
pyjwt
flask-cors
gunicorn